requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
multidict==5.1.0
mysql-connector-python==8.0.22
oauthlib==3.1.0
protobuf==3.14.0
py-trello==0.17.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
